requirements.txt
python-dotenv
fastapi
pydantic
revChatGPT
uvicorn
EdgeGPT